Refactor the combo handler into a reusable Express route middleware.
In addition to the combo handler middleware, the 'server' lib can be
used to generate (or augment) an Express server with a set of
configured combo routes.

app.js is now just a convenience script for running an instance of the
barebones server based on the settings in config.js.